move LazyHDF5Matrix to vibra/engine/serialization
