+++ vibra/engine/serialization/project_writer.py
@@ -20,9 +20,9 @@ from vibra.engine.properties.libraries.material_library import MaterialLibrary
 from vibra.engine.properties.material import Material
 from vibra.engine.properties.model_properties import ModelProperties
 from vibra.engine.serialization.file_helpers import read_json, update_json, write_image, write_json
+from vibra.engine.serialization.lazy_hdf5_matrix import LazyHDF5MatrixWriter
 from vibra.engine.solution import HarmonicSolution, ModalSolution
 from vibra.engine.solution.lazy_harmonic_solution import LazyHarmonicSolution
-from vibra.project_files.lazy_hdf5_matrix import LazyHDF5MatrixWriter
 
 from .project_hasher import HashEnum, ProjectHasher
 from .project_paths import ProjectPaths
